+++ hub20/apps/raiden/client/node.py
@@ -89,6 +89,10 @@ class RaidenClient:
     def raiden_token_list_endpoint(self) -> str:
         return f"{self.raiden_root_endpoint}/tokens"
 
+    @property
+    def raiden_udc_endpoint(self) -> str:
+        return f"{self.raiden_root_endpoint}/user_deposit"
+
     def channel_endpoint(self, channel: Channel) -> str:
         raiden_endpoint = self.raiden_root_endpoint
         return f"{raiden_endpoint}/channels/{channel.token.address}/{channel.partner_address}"
@@ -131,9 +135,10 @@ class RaidenClient:
         except RaidenConnectionError:
             return "offline"
 
-    def join_token_network(self, token_network: TokenNetwork, amount: EthereumTokenAmount):
-        url = self.token_network_endpoint(token_network=token_network)
-        return _make_request(url, method="PUT", funds=amount.as_wei)
+    def make_user_deposit(self, total_deposit_amount: EthereumTokenAmount):
+        return _make_request(
+            self.raiden_udc_endpoint, method="POST", total_deposit=total_deposit_amount.as_wei
+        )
 
     def leave_token_network(self, token_network: TokenNetwork):
         url = self.token_network_endpoint(token_network=token_network)