+++ CHANGELOG.md
@@ -5,11 +5,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+ - API Endpoint to list Token Networks, optionally filtered by chain
+   and/or if we have a channel connection to it.
+
+
 ### Changed
 
  - Refactored transfer model, broken from single Transfer class into
    Internal/Blockchain/Raiden transfers, and corresponding
    execution/confirmation models
+
+ - Raiden operations (channel funding, User Deposit funding,
+   joining/leaving token network) are done through the Raiden Node and
+   the celery tasks are tracked. The results are stored on the
+   database.
+
 
 ## [0.4.0] - 2022-01-31